requirements/dev.txt
django-debug-toolbar==3.2.*
django-extensions==2.2.*
flake8==3.9.*
autopep8==1.5.*
selenium==4.0.*
bs4==1.5.*
oauth2client==1.5.*